Collect list components into one file
