--- packages/changed-elements-react/src/widgets/comparisonJobWidget/components/VersionCompareSelectorInner.tsx
@@ -1,81 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-import { IModelApp } from "@itwin/core-frontend";
-import { Text } from "@itwin/itwinui-react";
-import { ReactNode } from "react";
-
-import type { NamedVersion } from "../../../clients/iModelsClient";
-import type { VersionState } from "../NamedVersions.js";
-import { ManageNamedVersions } from "./VersionCompareManageNamedVersions";
-import { CurrentVersionEntry } from "./VersionEntries";
-import { VersionList } from "./VersionList";
-
-interface VersionCompareSelectorInnerProps {
-  entries: NamedVersion[];
-  versionState: VersionState[];
-  currentVersion: NamedVersion | undefined;
-  selectedVersionChangesetId: string | undefined;
-  onVersionClicked: (targetVersion: NamedVersion) => void;
-  wantTitle: boolean | undefined;
-
-  /** Optional prop for a user supplied component to handle managing named versions.*/
-  manageNamedVersionsSlot?: ReactNode | undefined;
-
-  /** If true display loading spinner to indicate we are receiving more named versions*/
-  isLoading: boolean;
-}
-
-/**
- * Component that houses named version list.
- * Also displays the current versions information.
- */
-export function VersionCompareSelectorInner(props: VersionCompareSelectorInnerProps) {
-  return (
-    <div className="comparison-job-selector">
-      {
-        props.currentVersion &&
-        <div className="comparison-job-row">
-          <div className="current-comparison-title">
-            {`${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.compare")}:`}
-          </div>
-          <div className="comparison-job-container-current">
-            <CurrentVersionEntry namedVersion={props.currentVersion} isProcessed />
-          </div>
-        </div>
-      }
-      {
-        props.wantTitle &&
-        <div className="title">
-          {IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.versionCompare")}
-        </div>
-      }
-      {<div className="comparison-job-label">
-        {`${IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.withPrevious")}:`}
-      </div>}
-      {
-        props.entries.length > 0 && props.currentVersion ? (
-          <VersionList
-            entries={props.entries}
-            versionState={props.versionState}
-            currentVersion={props.currentVersion}
-            selectedVersionChangesetId={props.selectedVersionChangesetId}
-            onVersionClicked={props.onVersionClicked}
-            isLoading={props.isLoading}
-          />
-        ) : (
-          <Text className="no-named-versions-message" variant="leading">
-            {IModelApp.localization.getLocalizedString("VersionCompare:versionCompare.noPastNamedVersions")}
-          </Text>
-        )
-      }
-      {
-        props.manageNamedVersionsSlot &&
-        <ManageNamedVersions>
-          {props.manageNamedVersionsSlot}
-        </ManageNamedVersions>
-      }
-    </div>
-  );
-}
